server/db_migration_service/requirements.txt
psycopg2-binary~=2.9.7
common-lib~=0.1.0